native/scripts/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
supabase
chromadb
flask
